Move QueryForm into searchpromotions
