+++ wagtail/contrib/search_promotions/forms.py
@@ -6,6 +6,18 @@ from wagtail.admin.widgets import AdminPageChooser
 from wagtail.contrib.search_promotions.models import Query, SearchPromotion
 
 
+class QueryForm(forms.Form):
+    query_string = forms.CharField(
+        label=_("Search term(s)/phrase"),
+        help_text=_(
+            "Enter the full search string to match. An "
+            "exact match is required for your Promoted Results to be "
+            "displayed, wildcards are NOT allowed."
+        ),
+        required=True,
+    )
+
+
 class SearchPromotionForm(forms.ModelForm):
     sort_order = forms.IntegerField(required=False)
