+++ wagtail/contrib/search_promotions/views.py
@@ -13,7 +13,6 @@ from wagtail.admin.forms.search import SearchForm
 from wagtail.contrib.search_promotions import forms
 from wagtail.contrib.search_promotions.models import Query
 from wagtail.log_actions import log
-from wagtail.search import forms as search_forms
 
 
 @any_permission_required(
@@ -124,7 +123,7 @@ def save_searchpicks(query, new_query, searchpicks_formset):
 def add(request):
     if request.method == "POST":
         # Get query
-        query_form = search_forms.QueryForm(request.POST)
+        query_form = forms.QueryForm(request.POST)
         if query_form.is_valid():
             query = Query.get(query_form["query_string"].value())
 
@@ -164,7 +163,7 @@ def add(request):
         else:
             searchpicks_formset = forms.SearchPromotionsFormSet()
     else:
-        query_form = search_forms.QueryForm()
+        query_form = forms.QueryForm()
         searchpicks_formset = forms.SearchPromotionsFormSet()
 
     return TemplateResponse(
@@ -184,7 +183,7 @@ def edit(request, query_id):
 
     if request.method == "POST":
         # Get query
-        query_form = search_forms.QueryForm(request.POST)
+        query_form = forms.QueryForm(request.POST)
         # and the recommendations
         searchpicks_formset = forms.SearchPromotionsFormSet(
             request.POST, instance=query
@@ -222,9 +221,7 @@ def edit(request, query_id):
                     # specific errors will be displayed within form fields
 
     else:
-        query_form = search_forms.QueryForm(
-            initial={"query_string": query.query_string}
-        )
+        query_form = forms.QueryForm(initial={"query_string": query.query_string})
         searchpicks_formset = forms.SearchPromotionsFormSet(instance=query)
 
     return TemplateResponse(
