+++ wagtail/search/forms.py
@@ -1,14 +1,13 @@
-from django import forms
-from django.utils.translation import gettext_lazy as _
+import sys
 
+from wagtail.utils.deprecation import MovedDefinitionHandler, RemovedInWagtail60Warning
 
-class QueryForm(forms.Form):
-    query_string = forms.CharField(
-        label=_("Search term(s)/phrase"),
-        help_text=_(
-            "Enter the full search string to match. An "
-            "exact match is required for your Promoted Results to be "
-            "displayed, wildcards are NOT allowed."
-        ),
-        required=True,
-    )
+MOVED_DEFINITIONS = {
+    "QueryForm": ("wagtail.contrib.search_promotions.forms", "QueryForm"),
+}
+
+sys.modules[__name__] = MovedDefinitionHandler(
+    sys.modules[__name__],
+    MOVED_DEFINITIONS,
+    RemovedInWagtail60Warning,
+)
